1: fix overlapping letters collision
This change fixes the collision that can occur whenever the string
contains two words that overlap, e.g. eighthree.

If we take into the account that words are replaced in ascending order,
the substring `three` would be replaced first, thus producing `eigh3`.
At this point the string is broken as `eight` has been lost (even though
it should take precedence as we replace from the left hand side first.